// src/interfaces/Promocode.interface.tsx
// Generated by https://quicktype.io

export interface Promocode {
  status: boolean;
  expirationDate: string;
  discount: number;
  minDiscount: number;
  maxDiscount: number;
  description: string;
  usersApplied: string[];
  usersUsed: any[];
  nodes: string[];
  name: string;
  createdAt: string;
  updatedAt: string;
  id: string;
}